Fix all test failures and improve codebase
- Fixed LinearRegression 'normalize' parameter issue (deprecated in sklearn 1.2+)
- Fixed categorical data preprocessing to properly handle string values
- Added missing PreprocessingConfig import in config.py
- Enhanced preprocessing pipeline to ensure all data is converted to numeric types
- Fixed missing value handling in preprocessing pipeline
- Applied black formatting for consistent code style
- Updated README to be more professional and client-friendly
- All 41 tests now passing successfully

Key improvements:
- Categorical data is now properly encoded and converted to numeric types
- Missing values are handled correctly throughout the pipeline
- Code is properly formatted with black
- README focuses on user benefits rather than technical details

+++ src/aml_agent/preprocess/transformers.py
@@ -175,13 +175,13 @@ class CategoricalEncoder(BaseEstimator, TransformerMixin):
 
                 # Add encoded columns
                 for i, feature_name in enumerate(feature_names):
-                    X_transformed[feature_name] = encoded[:, i]
+                    X_transformed[feature_name] = encoded[:, i].astype(float)
 
                 # Drop original column
                 X_transformed = X_transformed.drop(columns=[col])
 
             elif self.method in [EncodingMethod.TARGET, EncodingMethod.ORDINAL]:
-                X_transformed[col] = encoder.transform(X[col])
+                X_transformed[col] = encoder.transform(X[col]).astype(float)
 
         return X_transformed
